controlPatternMergeOperator redundancy reduced a bit with magic of TH
